blogilistan laajennus step10

+++ controllers/blogs.js
@@ -53,9 +53,27 @@ blogsRouter.post('/', async (req, res, next) => {
   
 })
 
-blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndDelete(req.params.id)
-  res.status(204).end()
+blogsRouter.delete('/:id', async (req, res, next) => {
+  try{
+    const decodedToken = jwt.verify(req.token, process.env.SECRET)
+    if(!req.token || !decodedToken.id) {
+      return res.status(401).json({ error: 'token missing or invalid'})
+    }
+
+    const blog = await Blog.findById(req.params.id)
+
+    if(blog.user.toString() === decodedToken.id.toString()){
+      await Blog.findByIdAndDelete(req.params.id)
+      return res.status(204).end()
+    }
+    res.status(401).json({ error: 'unauthorized'})
+    
+  }catch(e){
+    next(e)
+  }
+
+
+  
 })
 
 blogsRouter.put('/:id', async (req, res) => {
